feat: typescript hokey pokey
