+++ packages/api-client-modal/package.json
@@ -5,7 +5,6 @@
   "author": "Scalar (https://github.com/scalar)",
   "homepage": "https://github.com/scalar/scalar",
   "bugs": "https://github.com/scalar/scalar/issues/new/choose",
-  "private": true,
   "repository": {
     "type": "git",
     "url": "https://github.com/scalar/scalar.git",
@@ -15,6 +14,7 @@
     "vue api client modal"
   ],
   "version": "0.0.1",
+  "private": true,
   "engines": {
     "node": ">=18"
   },
