+++ packages/api-client-modal/playground/main.ts
@@ -5,6 +5,9 @@ import content from '@scalar/galaxy/latest.json?raw'
 const el = document.getElementById('root')
 console.log(el)
 
-const { open } = await createScalarClient(el, {
-  spec: { content },
-})
+if (el) {
+  const { modalState } = await createScalarClient(el, {
+    spec: { content },
+  })
+  modalState.show()
+}
