+++ packages/api-reference/src/components/Content/Introduction/Introduction.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
+import type { Spec } from '@scalar/oas-utils'
 import type { OpenAPIV2, OpenAPIV3, OpenAPIV3_1 } from '@scalar/openapi-parser'
 import GithubSlugger from 'github-slugger'
 import { computed } from 'vue'
 
-import type { Spec } from '../../../types'
 import { Badge } from '../../Badge'
 import {
   Section,
