+++ packages/api-reference/src/helpers/createEmptySpecification.ts
@@ -1,4 +1,4 @@
-import type { Spec } from '../types'
+import type { Spec } from '@scalar/oas-utils'
 
 export function createEmptySpecification(): Spec {
   return {
