+++ packages/api-reference/src/helpers/hasSecuritySchemes.ts
@@ -1,4 +1,4 @@
-import type { Spec } from '../types'
+import type { Spec } from '@scalar/oas-utils'
 
 export const hasSecuritySchemes = (spec?: Spec) => {
   // TODO: Show security schemes
