+++ packages/api-reference/src/hooks/useSidebar.ts
@@ -1,5 +1,5 @@
 import { useApiClientStore, useOpenApiStore } from '@scalar/api-client'
-import type { TransformedOperation } from '@scalar/oas-utils'
+import type { Spec, Tag, TransformedOperation } from '@scalar/oas-utils'
 import { ssrState } from '@scalar/oas-utils/helpers'
 import type { OpenAPIV3_1 } from '@scalar/openapi-parser'
 import { computed, reactive, ref, watch } from 'vue'
@@ -12,7 +12,6 @@ import {
   hasWebhooks,
   openClientFor,
 } from '../helpers'
-import type { Spec, Tag } from '../types'
 import { useNavState } from './useNavState'
 
 export type SidebarEntry = {
