+++ packages/api-reference/src/types.ts
@@ -1,11 +1,9 @@
-import type { Server } from '@scalar/api-client'
 import type {
   AuthenticationState,
   ContentType,
+  Spec,
   SpecConfiguration,
-  TransformedOperation,
 } from '@scalar/oas-utils'
-import type { OpenAPIV2, OpenAPIV3, OpenAPIV3_1 } from '@scalar/openapi-parser'
 import type { ThemeId } from '@scalar/themes'
 import type { UseSeoMetaInput } from '@unhead/schema'
 import type { HarRequest, TargetId } from 'httpsnippet-lite'
@@ -167,12 +165,6 @@ export type ExampleResponseHeaders = Record<
   }
 >
 
-export type Tag = {
-  'name': string
-  'description': string
-  'operations': TransformedOperation[]
-  'x-displayName'?: string
-}
 export type Parameter = {
   name: string
   required: boolean
@@ -220,54 +212,6 @@ export type Info = {
   contact?: Contact
   license?: License
   version?: string
-}
-
-export type ExternalDocs = {
-  description: string
-  url: string
-}
-
-export type TagGroup = {
-  name: string
-  tags: string[]
-}
-
-export type SecurityScheme =
-  | Record<string, never> // Empty objects
-  | OpenAPIV2.SecuritySchemeObject
-  | OpenAPIV3.SecuritySchemeObject
-  | OpenAPIV3_1.SecuritySchemeObject
-
-export type Definitions = OpenAPIV2.DefinitionsObject
-
-export type Webhooks = Record<
-  string,
-  Record<
-    OpenAPIV3_1.HttpMethods,
-    TransformedOperation & {
-      'x-internal'?: boolean
-    }
-  >
->
-
-export type Spec = {
-  'tags'?: Tag[]
-  'info':
-    | Partial<OpenAPIV2.Document['info']>
-    | Partial<OpenAPIV3.Document['info']>
-    | Partial<OpenAPIV3_1.Document['info']>
-  'host'?: OpenAPIV2.Document['host']
-  'basePath'?: OpenAPIV2.Document['basePath']
-  'schemes'?: OpenAPIV2.Document['schemes']
-  'externalDocs'?: ExternalDocs
-  'servers'?: Server[]
-  'components'?: OpenAPIV3.ComponentsObject | OpenAPIV3_1.ComponentsObject
-  'webhooks'?: Webhooks
-  'definitions'?: Definitions
-  'swagger'?: OpenAPIV2.Document['swagger']
-  'openapi'?: OpenAPIV3.Document['openapi'] | OpenAPIV3_1.Document['openapi']
-  'x-tagGroups'?: TagGroup[]
-  'security'?: OpenAPIV3.SecurityRequirementObject[]
 }
 
 export type HarRequestWithPath = HarRequest & {
