+++ packages/client-app/src/components/ActionModal/ActionModalImport.vue
@@ -24,7 +24,7 @@ const { open: openSpecFileDialog } = useFileDialog({
       const reader = new FileReader()
       reader.onload = async (e) => {
         const text = e.target?.result as string
-        importSpecFile(text)
+        importSpecFile({ spec: text })
         handleSubmit()
       }
       reader.readAsText(file)
