+++ packages/oas-utils/package.json
@@ -105,6 +105,7 @@
     "openapi-types": "^12.1.3",
     "rollup": "^4.16.4",
     "tsc-alias": "^1.8.8",
+    "type-fest": "^4.20.0",
     "vite": "^5.2.10",
     "vitest": "^1.6.0"
   },
