+++ packages/oas-utils/src/transforms/import-spec.ts
@@ -6,9 +6,10 @@ import { tagObjectSchema } from '@/entities/workspace/spec/spec'
 import type { RequestMethod } from '@/helpers'
 import { parseJsonOrYaml } from '@/helpers/parse'
 import { schemaModel } from '@/helpers/schema-model'
-import type { AnyObject } from '@/types'
+import type { AnyObject, Spec } from '@/types'
 import { openapi } from '@scalar/openapi-parser'
 import type { OpenAPIV3_1 } from 'openapi-types'
+import type { RequireExactlyOne } from 'type-fest'
 
 const PARAM_DICTIONARY = {
   cookie: 'cookies',
@@ -17,11 +18,21 @@ const PARAM_DICTIONARY = {
   query: 'query',
 } as const
 
+export type ImportSpecPayload = RequireExactlyOne<
+  {
+    spec: string | AnyObject
+    parsedSpec: Spec
+  },
+  'spec' | 'parsedSpec'
+>
+
 /** Import an OpenAPI spec file and convert it to workspace entities */
-export const importSpecToWorkspace = async (spec: string | AnyObject) => {
+export const importSpecToWorkspace = async (payload: ImportSpecPayload) => {
   const importWarnings: string[] = []
   const requests: Request[] = []
-  const parsedSpec = parseJsonOrYaml(spec) as OpenAPIV3_1.Document
+  const parsedSpec =
+    payload.parsedSpec ??
+    (parseJsonOrYaml(payload.spec) as OpenAPIV3_1.Document)
 
   const { schema, errors } = await openapi().load(parsedSpec).resolve()
   if (errors?.length || !schema) {
