+++ packages/play-button/package.json
@@ -52,6 +52,7 @@
   "dependencies": {
     "@scalar/api-client": "workspace:*",
     "@scalar/api-reference": "workspace:*",
+    "@scalar/oas-utils": "workspace:*",
     "@scalar/openapi-parser": "^0.3.2",
     "@scalar/themes": "workspace:*",
     "@vueuse/core": "^10.10.0"
