+++ packages/play-button/src/index.ts
@@ -2,12 +2,8 @@
  * This file is the entry point for the CDN version of the Scalar Test Button.
  * It’s responsible for finding the spec and configuration in the HTML, and mounting the Vue.js app.
  */
-import {
-  ApiClientModal,
-  type Spec,
-  openClientFor,
-  parse,
-} from '@scalar/api-reference'
+import { ApiClientModal, openClientFor, parse } from '@scalar/api-reference'
+import type { Spec } from '@scalar/oas-utils'
 import { createApp, h, reactive } from 'vue'
 
 const specScriptTag = document.getElementById('scalar-play-button-script')
